fixed Trends component, fixed func inside Tweets
